authenticate endpoint and disable login

+++ src/main/java/qadr/bank/api/service/BankService.java
@@ -1,15 +1,16 @@
 package qadr.bank.api.service;
 
+import org.springframework.security.core.userdetails.User;
 import qadr.bank.api.model.Bank;
 
 import java.util.List;
 
 public interface BankService {
     Bank addBank(Bank bank);
-    Bank getBank(long id);
+    Bank getBank(int id);
     Bank getBank (String name);
     List<Bank> getBanks(String type);
     List<Bank> getAll();
-    Bank deleteBank(long id);
-    Bank updateBank(long id, Bank bank);
+    Bank deleteBank(int id);
+    Bank updateBank(int id, Bank bank);
 }
