+++ src/main/java/qadr/bank/api/service/BankServiceImpl.java
@@ -2,6 +2,10 @@ package qadr.bank.api.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import qadr.bank.api.errors.CustomException;
@@ -25,7 +29,7 @@ public class BankServiceImpl implements BankService {
     }
 
     @Override
-    public Bank getBank(long id) {
+    public Bank getBank(int id) {
         Optional<Bank> bankOptional =  bankRepo.findById(id);
         if (bankOptional.isEmpty()) throw new CustomException("Bank not found", HttpStatus.NOT_FOUND);
         return bankOptional.get();
@@ -48,7 +52,7 @@ public class BankServiceImpl implements BankService {
     }
 
     @Override
-    public Bank deleteBank(long id) {
+    public Bank deleteBank(int id) {
         Bank bank = bankRepo.findById(id)
                 .orElseThrow(
                         ()-> new CustomException("Bank not found", HttpStatus.NOT_FOUND)
@@ -58,7 +62,7 @@ public class BankServiceImpl implements BankService {
     }
 
     @Override
-    public Bank updateBank(long id, Bank newBank) {
+    public Bank updateBank(int id, Bank newBank) {
         Bank bank = bankRepo.findById(id).orElseThrow(
                 () -> new CustomException("Bank not found", HttpStatus.NOT_FOUND)
         );
@@ -74,5 +78,6 @@ public class BankServiceImpl implements BankService {
         bank.setSortCode(newBank.getSortCode());
         return bank;
     }
+
 
 }
